Rename `sequences` to `auton`
