+++ src/main/java/frc/robot/RobotContainer.java
@@ -16,15 +16,15 @@ import edu.wpi.first.wpilibj2.command.RunCommand;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 
 import frc.robot.commands.BalancerCommand;
+import frc.robot.commands.auton.BalanceAutonSequence;
+import frc.robot.commands.auton.BottomAutonSequence;
+import frc.robot.commands.auton.TopAutonSequence;
+import frc.robot.commands.auton.test.BoxAutonSequence;
+import frc.robot.commands.auton.test.GRTAutonSequence;
+import frc.robot.commands.auton.test.HighRotationLinePath;
+import frc.robot.commands.auton.test.RotatingSCurveAutonSequence;
+import frc.robot.commands.auton.test.StraightLinePath;
 import frc.robot.commands.dropping.DropperChooserCommand;
-import frc.robot.commands.sequences.BalanceAutonSequence;
-import frc.robot.commands.sequences.BottomAutonSequence;
-import frc.robot.commands.sequences.TopAutonSequence;
-import frc.robot.commands.sequences.test.BoxAutonSequence;
-import frc.robot.commands.sequences.test.GRTAutonSequence;
-import frc.robot.commands.sequences.test.HighRotationLinePath;
-import frc.robot.commands.sequences.test.RotatingSCurveAutonSequence;
-import frc.robot.commands.sequences.test.StraightLinePath;
 import frc.robot.controllers.BaseDriveController;
 import frc.robot.controllers.DualJoystickDriveController;
 import frc.robot.controllers.TwistJoystickDriveController;
