+++ src/main/java/frc/robot/commands/auton/BaseAutonSequence.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.sequences;
+package frc.robot.commands.auton;
 
 import java.util.List;
 
@@ -6,7 +6,6 @@ import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import edu.wpi.first.wpilibj2.command.WaitCommand;
 
 import frc.robot.commands.dropping.DropperChooserCommand;
 import frc.robot.commands.grabber.RollerIntakeCommand;
